Don't hardcode URL scheme, domain, port in HTML
Serving hardcoded HTML base URL scheme, domain and port from the server's
point of view to the client breaks when proxying to use publicly facing
settings different from the one configured internally in the webserver.

The context path is still used, and needs to be the same internally as
well as externally.

This enables for example proxied HTTPS support, while allowing export services
to make requests which do not require roundtrips outside of the server.